# modules/claims_api/app/swagger/claims_api/v2/dev/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Benefits Claims",
    "version": "v2",
...
  },
...
  "components": {
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "productionOauth": {
        "type": "oauth2",
        "description": "This API uses OAuth 2 with the authorization code grant flow. [More info](https://developer.va.gov/explore/authorization?api=claims)",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.va.gov/oauth2/authorization",
            "tokenUrl": "https://api.va.gov/oauth2/token",
            "scopes": {
              "claim.read": "Retrieve claim data",
              "claim.write": "Submit claim data"
            }
          }
        }
      },
      "sandboxOauth": {
        "type": "oauth2",
        "description": "This API uses OAuth 2 with the authorization code grant flow. [More info](https://developer.va.gov/explore/authorization?api=claims)",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sandbox-api.va.gov/oauth2/authorization",
            "tokenUrl": "https://sandbox-api.va.gov/oauth2/token",
            "scopes": {
              "claim.read": "Retrieve claim data",
              "claim.write": "Submit claim data"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/veterans/{veteranId}/claims": {
      "get": {
        "summary": "Find all benefits claims for a Veteran.",
        "tags": [
          "Claims"
        ],
        "operationId": "findClaims",
        "security": [
          {
            "productionOauth": [
              "claim.read"
            ]
          },
          {
            "sandboxOauth": [
              "claim.read"
            ]
          },
          {
            "bearer_token": [

            ]
          }
        ],
        "description": "Retrieves all claims for Veteran.",
        "parameters": [
          {
            "name": "veteranId",
            "in": "path",
            "required": true,
            "example": "1012667145V762142",
            "description": "ID of Veteran",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "claim response",
            "content": {
              "application/json": {
                "example": [
                  {
                    "claimDate": "2017-05-02",
                    "claimId": "555555555",
                    "claimPhaseDates": {
                      "phaseChangeDate": "2017-10-18"
                    },
                    "claimType": "Compensation",
                    "closeDate": "2017-10-18",
                    "decisionLetterSent": false,
                    "developmentLetterSent": false,
                    "documentsNeeded": false,
                    "endProductCode": "404",
                    "evidenceWaiverSubmitted5103": false,
                    "lighthouseId": null,
                    "status": "PENDING"
                  }
                ],
                "schema": {
                  "type": "array",
                  "items": {
                    "additionalProperties": false,
                    "description": "Claim details",
                    "properties": {
                      "claimId": {
                        "type": "string",
                        "nullable": true,
                        "description": "Claim ID in VBMS",
                        "example": "600131328"
                      },
                      "claimType": {
                        "type": "string",
                        "description": "Name of claim type",
                        "example": "Compensation"
                      },
                      "claimDate": {
                        "format": "date",
                        "type": "string",
                        "description": "Date in YYYY-MM-DD the claim was first filed",
                        "example": "2018-06-04"
                      },
                      "claimPhaseDates": {
                        "type": "object",
                        "properties": {
                          "phaseChangeDate": {
                            "format": "date",
                            "type": "string",
                            "description": "The date that the claim changed to its current phase",
                            "example": "2017-10-18"
                          }
                        }
                      },
                      "closeDate": {
                        "format": "date",
                        "type": "string",
                        "description": "Date claim was closed",
                        "example": "2019-09-04"
                      },
                      "developmentLetterSent": {
                        "type": "boolean",
                        "description": "If true, a development letter has been sent to the claimant regarding a benefit claim",
                        "example": "false"
                      },
                      "decisionLetterSent": {
                        "type": "boolean",
                        "description": "If true, a decision letter has been sent to the claimant regarding a benefit claim",
                        "example": "false"
                      },
                      "documentsNeeded": {
                        "type": "boolean",
                        "description": "If true, the claim requires additional documents to be submitted",
                        "example": "false"
                      },
                      "endProductCode": {
                        "type": "string",
                        "description": "End product code of claim"
                      },
                      "evidenceWaiverSubmitted5103": {
                        "type": "boolean",
                        "description": "If true, indicates a decision has been requested and/or a Waiver 5103 has been submitted",
                        "example": "false"
                      },
                      "lighthouseId": {
                        "type": "string",
                        "nullable": true,
                        "description": "Claim ID in Lighthouse",
                        "example": "0BAEFC26-1CE4-4046-9B3C-3071055603DB"
                      },
                      "status": {
                        "type": "string",
                        "description": "Status of claim",
                        "enum": [
                          "PENDING",
                          "CLAIM_RECEIVED",
                          "INITIAL_REVIEW",
                          "EVIDENCE_GATHERING_REVIEW_DECISION",
                          "PREPARATION_FOR_NOTIFICATION",
                          "COMPLETE",
                          "ERRORED"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "errors": [
                    {
                      "title": "Not authorized",
                      "detail": "Not authorized"
                    }
                  ]
                },
                "schema": {
                  "required": [
                    "errors"
                  ],
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "required": [
                          "title",
                          "detail"
                        ],
                        "properties": {
                          "title": {
                            "type": "string",
                            "description": "HTTP error title"
                          },
                          "detail": {
                            "type": "string",
                            "description": "HTTP error detail"
                          },
                          "source": {
                            "type": "object",
                            "additionalProperties": false,
                            "description": "Source of error",
                            "properties": {
                              "pointer": {
                                "type": "string",
                                "description": "Pointer to source of error"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "errors": [
                    {
                      "title": "Forbidden",
                      "detail": "Forbidden"
                    }
                  ]
                },
                "schema": {
                  "required": [
                    "errors"
                  ],
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "required": [
                          "title",
                          "detail"
                        ],
                        "properties": {
                          "title": {
                            "type": "string",
                            "description": "HTTP error title"
                          },
                          "detail": {
                            "type": "string",
                            "description": "HTTP error detail"
                          },
                          "source": {
                            "type": "object",
                            "additionalProperties": false,
                            "description": "Source of error",
                            "properties": {
                              "pointer": {
                                "type": "string",
                                "description": "Pointer to source of error"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://sandbox-api.va.gov/services/claims/{version}",
      "description": "VA.gov API sandbox environment",
      "variables": {
        "version": {
          "default": "v2"
        }
      }
    },
    {
      "url": "https://api.va.gov/services/claims/{version}",
      "description": "VA.gov API production environment",
      "variables": {
        "version": {
          "default": "v2"
        }
      }
    }
  ]
}